Preparatory refactor: give the public page list a module of its own
`src/lib/seo.ts` imports the club's logo and hero asset manifests, which means
anything that wants the list of public pages has to be able to load those too.
The old screenshot script could (bun), a Playwright spec cannot.

So the list moves to `src/lib/public-pages.ts`, which imports nothing at all,
and seo.ts re-exports it: every existing import of `@/lib/seo` is unchanged.
The hand-written list of public pages that are deliberately noindex — and so
deliberately absent from the sitemap — comes with it, out of the screenshot
script and into source where a test can check it: each one has to be a real
route, has to actually set `robots: noindex`, and cannot also be in
PUBLIC_PAGES.

No behaviour change.
